requirements.txt
azure-functions
pandas
numpy
openpyxl
xlsxwriter